+++ tools/lib/generateJSONSchema.js
@@ -443,6 +443,7 @@ const generateJSONSchema = (sport /*: string*/) => {
 				type: "array",
 				items: {
 					oneOf: [
+						gameAttribute("aiJerseyRetirement", { type: "boolean" }),
 						gameAttribute("aiTradesFactor", { type: "number" }),
 						gameAttribute("allStarGame", { type: "boolean" }),
 						gameAttribute("autoDeleteOldBoxScores", { type: "boolean" }),